add improved pydantic schema for order processing

+++ app/database/models/product_model.py
@@ -21,10 +21,10 @@
 			=> single/multiple entry for its every version in productvariant
 """
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from sqlalchemy import String, Integer, ForeignKey, CheckConstraint
+from sqlalchemy import String, Integer, ForeignKey, CheckConstraint, Boolean, Text
 from typing import List
 from datetime import datetime
-from app.database import Base
+from app.database.db import Base
 
 class Products(Base):
 	__tablename__ = "products"
@@ -35,7 +35,7 @@ class Products(Base):
 	product_name: Mapped[str] = mapped_column(String(100), nullable=False)
 	short_desc: Mapped[str] = mapped_column(Text(255), nullable=False)
 	image_link: Mapped[str] = mapped_column(String(200), nullable=False)
-	current_price: Mapped[int] = mapped_column(Integer(100), nullable=False)
+	current_price: Mapped[int] = mapped_column(Integer, nullable=False)
 	in_stock: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False)
 
 	catg_id: Mapped[int] = mapped_column(ForeignKey(
